+++ tests/test_db_user_model.py
@@ -4,9 +4,9 @@ from datetime import datetime
 
 import pytest
 
-import utils.db_test_utils as db_utils
 from models.user import User
 from repositories.user_db_repository import UserDBRepository
+import utils.db_test_utils as db_utils
 from utils.auth_utils import AuthenticationUtils
