+++ utils/auth_utils.py
@@ -5,10 +5,10 @@ JWT token creation, and token verification. It uses bcrypt for password hashing
 and jose for JWT operations.
 """
 
-import logging
-import os
 from datetime import datetime, timedelta
 from functools import lru_cache
+import logging
+import os
 from typing import Optional
 
 import bcrypt
